Split Assert.ThrowsException.cs into focused partial files (#9090)
