+++ src/TestFramework/TestFramework/Assertions/Assert.ThrowsException.cs
@@ -1,8 +1,6 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using System.ComponentModel;
-
 namespace Microsoft.VisualStudio.TestTools.UnitTesting;
 
 #pragma warning disable RS0026 // Do not add multiple public overloads with optional parameters
@@ -14,268 +12,6 @@ namespace Microsoft.VisualStudio.TestTools.UnitTesting;
 /// </summary>
 public sealed partial class Assert
 {
-    /// <summary>
-    /// Provides an interpolated string handler used by <c>Assert.Throws</c> overloads
-    /// that only allocates and formats the message when the assertion is failing.
-    /// </summary>
-    /// <typeparam name="TException">The type of exception expected to be thrown.</typeparam>
-    /// <remarks>
-    /// This type is intended to be used by the compiler; users should not reference it directly.
-    /// </remarks>
-    [StackTraceHidden]
-    [InterpolatedStringHandler]
-    [EditorBrowsable(EditorBrowsableState.Never)]
-    public readonly struct AssertNonStrictThrowsInterpolatedStringHandler<TException>
-        where TException : Exception
-    {
-        private readonly StringBuilder? _builder;
-        private readonly ThrowsExceptionState _state;
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="AssertNonStrictThrowsInterpolatedStringHandler{TException}"/> struct.
-        /// </summary>
-        /// <param name="literalLength">The number of constant characters in the interpolated string.</param>
-        /// <param name="formattedCount">The number of interpolation expressions in the interpolated string.</param>
-        /// <param name="action">The delegate being asserted; the message is only computed when the assertion fails.</param>
-        /// <param name="shouldAppend">When this method returns, indicates whether the interpolated string should be evaluated.</param>
-        public AssertNonStrictThrowsInterpolatedStringHandler(int literalLength, int formattedCount, Action action, out bool shouldAppend)
-        {
-            _state = IsThrowsFailing<TException>(action, isStrictType: false);
-            shouldAppend = _state.FailureKind != ThrowsFailureKind.NotFailing;
-            if (shouldAppend)
-            {
-                _builder = new StringBuilder(literalLength + formattedCount);
-            }
-        }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="AssertNonStrictThrowsInterpolatedStringHandler{TException}"/> struct.
-        /// </summary>
-        /// <param name="literalLength">The number of constant characters in the interpolated string.</param>
-        /// <param name="formattedCount">The number of interpolation expressions in the interpolated string.</param>
-        /// <param name="action">The delegate being asserted; the message is only computed when the assertion fails.</param>
-        /// <param name="shouldAppend">When this method returns, indicates whether the interpolated string should be evaluated.</param>
-        public AssertNonStrictThrowsInterpolatedStringHandler(int literalLength, int formattedCount, Func<object?> action, out bool shouldAppend)
-            : this(literalLength, formattedCount, (Action)(() => _ = action()), out shouldAppend)
-        {
-        }
-
-        internal TException ComputeAssertion(string actionExpression)
-        {
-            if (_state.FailureKind != ThrowsFailureKind.NotFailing)
-            {
-                ReportThrowsFailed<TException>(isStrictType: false, _state, _builder!.ToString(), actionExpression, nameof(Throws));
-            }
-            else
-            {
-                return (TException)_state.ExceptionThrown!;
-            }
-
-            // Reached when ReportThrowsFailed records the failure into the active AssertScope and returns instead of throwing.
-            return null!;
-        }
-
-        /// <summary>Appends a literal string to the interpolated message.</summary>
-        /// <param name="value">The literal string to append.</param>
-        public void AppendLiteral(string value) => _builder!.Append(value);
-
-        /// <summary>Appends a formatted value to the interpolated message.</summary>
-        /// <typeparam name="T">The type of the value being appended.</typeparam>
-        /// <param name="value">The value to append.</param>
-        public void AppendFormatted<T>(T value) => AppendFormatted(value, format: null);
-
-#if NETCOREAPP3_1_OR_GREATER
-
-        /// <summary>Appends a formatted value to the interpolated message.</summary>
-        /// <param name="value">The value to append.</param>
-        public void AppendFormatted(ReadOnlySpan<char> value) => _builder!.Append(value);
-
-#pragma warning disable RS0027 // API with optional parameter(s) should have the most parameters amongst its public overloads
-
-        /// <summary>Appends a formatted value to the interpolated message.</summary>
-        /// <param name="value">The character span to append.</param>
-        /// <param name="alignment">The minimum width of the formatted value.</param>
-        /// <param name="format">The format string to use.</param>
-        public void AppendFormatted(ReadOnlySpan<char> value, int alignment = 0, string? format = null) => AppendFormatted(value.ToString(), alignment, format);
-#pragma warning restore RS0027 // API with optional parameter(s) should have the most parameters amongst its public overloads
-#endif
-
-        // NOTE: All the overloads involving format and/or alignment are not super efficient.
-        // This code path is only for when an assert is failing, so that's not the common scenario
-        // and should be okay if not very optimized.
-        // A more efficient implementation that can be used for .NET 6 and later is to delegate the work to
-        // the BCL's StringBuilder.AppendInterpolatedStringHandler
-
-        /// <summary>Appends a formatted value to the interpolated message.</summary>
-        /// <typeparam name="T">The type of the value being appended.</typeparam>
-        /// <param name="value">The value to append.</param>
-        /// <param name="format">The format string to use.</param>
-        public void AppendFormatted<T>(T value, string? format) => _builder!.AppendFormat(null, $"{{0:{format}}}", value);
-
-        /// <summary>Appends a formatted value to the interpolated message.</summary>
-        /// <typeparam name="T">The type of the value being appended.</typeparam>
-        /// <param name="value">The value to append.</param>
-        /// <param name="alignment">The minimum width of the formatted value.</param>
-        public void AppendFormatted<T>(T value, int alignment) => _builder!.AppendFormat(null, $"{{0,{alignment}}}", value);
-
-        /// <summary>Appends a formatted value to the interpolated message.</summary>
-        /// <typeparam name="T">The type of the value being appended.</typeparam>
-        /// <param name="value">The value to append.</param>
-        /// <param name="alignment">The minimum width of the formatted value.</param>
-        /// <param name="format">The format string to use.</param>
-        public void AppendFormatted<T>(T value, int alignment, string? format) => _builder!.AppendFormat(null, $"{{0,{alignment}:{format}}}", value);
-
-        /// <summary>Appends a formatted value to the interpolated message.</summary>
-        /// <param name="value">The value to append.</param>
-        public void AppendFormatted(string? value) => _builder!.Append(value);
-
-#pragma warning disable RS0027 // API with optional parameter(s) should have the most parameters amongst its public overloads
-
-        /// <summary>Appends a formatted value to the interpolated message.</summary>
-        /// <param name="value">The value to append.</param>
-        /// <param name="alignment">The minimum width of the formatted value.</param>
-        /// <param name="format">The format string to use.</param>
-        public void AppendFormatted(string? value, int alignment = 0, string? format = null) => _builder!.AppendFormat(null, $"{{0,{alignment}:{format}}}", value);
-
-        /// <summary>Appends a formatted value to the interpolated message.</summary>
-        /// <param name="value">The value to append.</param>
-        /// <param name="alignment">The minimum width of the formatted value.</param>
-        /// <param name="format">The format string to use.</param>
-        public void AppendFormatted(object? value, int alignment = 0, string? format = null) => _builder!.AppendFormat(null, $"{{0,{alignment}:{format}}}", value);
-#pragma warning restore RS0027 // API with optional parameter(s) should have the most parameters amongst its public overloads
-    }
-
-    /// <summary>
-    /// Provides an interpolated string handler used by <c>Assert.ThrowsExactly</c> overloads
-    /// that only allocates and formats the message when the assertion is failing.
-    /// </summary>
-    /// <typeparam name="TException">The type of exception expected to be thrown.</typeparam>
-    /// <remarks>
-    /// This type is intended to be used by the compiler; users should not reference it directly.
-    /// </remarks>
-    [StackTraceHidden]
-    [InterpolatedStringHandler]
-    [EditorBrowsable(EditorBrowsableState.Never)]
-    public readonly struct AssertThrowsExactlyInterpolatedStringHandler<TException>
-        where TException : Exception
-    {
-        private readonly StringBuilder? _builder;
-        private readonly ThrowsExceptionState _state;
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="AssertThrowsExactlyInterpolatedStringHandler{TException}"/> struct.
-        /// </summary>
-        /// <param name="literalLength">The number of constant characters in the interpolated string.</param>
-        /// <param name="formattedCount">The number of interpolation expressions in the interpolated string.</param>
-        /// <param name="action">The delegate being asserted; the message is only computed when the assertion fails.</param>
-        /// <param name="shouldAppend">When this method returns, indicates whether the interpolated string should be evaluated.</param>
-        public AssertThrowsExactlyInterpolatedStringHandler(int literalLength, int formattedCount, Action action, out bool shouldAppend)
-        {
-            _state = IsThrowsFailing<TException>(action, isStrictType: true);
-            shouldAppend = _state.FailureKind != ThrowsFailureKind.NotFailing;
-            if (shouldAppend)
-            {
-                _builder = new StringBuilder(literalLength + formattedCount);
-            }
-        }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="AssertThrowsExactlyInterpolatedStringHandler{TException}"/> struct.
-        /// </summary>
-        /// <param name="literalLength">The number of constant characters in the interpolated string.</param>
-        /// <param name="formattedCount">The number of interpolation expressions in the interpolated string.</param>
-        /// <param name="action">The delegate being asserted; the message is only computed when the assertion fails.</param>
-        /// <param name="shouldAppend">When this method returns, indicates whether the interpolated string should be evaluated.</param>
-        public AssertThrowsExactlyInterpolatedStringHandler(int literalLength, int formattedCount, Func<object?> action, out bool shouldAppend)
-            : this(literalLength, formattedCount, (Action)(() => _ = action()), out shouldAppend)
-        {
-        }
-
-        internal TException ComputeAssertion(string actionExpression)
-        {
-            if (_state.FailureKind != ThrowsFailureKind.NotFailing)
-            {
-                ReportThrowsFailed<TException>(isStrictType: true, _state, _builder!.ToString(), actionExpression, nameof(ThrowsExactly));
-            }
-            else
-            {
-                return (TException)_state.ExceptionThrown!;
-            }
-
-            // Reached when ReportThrowsFailed records the failure into the active AssertScope and returns instead of throwing.
-            return null!;
-        }
-
-        /// <summary>Appends a literal string to the interpolated message.</summary>
-        /// <param name="value">The literal string to append.</param>
-        public void AppendLiteral(string value) => _builder!.Append(value);
-
-        /// <summary>Appends a formatted value to the interpolated message.</summary>
-        /// <typeparam name="T">The type of the value being appended.</typeparam>
-        /// <param name="value">The value to append.</param>
-        public void AppendFormatted<T>(T value) => AppendFormatted(value, format: null);
-
-#if NETCOREAPP3_1_OR_GREATER
-
-        /// <summary>Appends a formatted value to the interpolated message.</summary>
-        /// <param name="value">The value to append.</param>
-        public void AppendFormatted(ReadOnlySpan<char> value) => _builder!.Append(value);
-
-#pragma warning disable RS0027 // API with optional parameter(s) should have the most parameters amongst its public overloads
-
-        /// <summary>Appends a formatted value to the interpolated message.</summary>
-        /// <param name="value">The character span to append.</param>
-        /// <param name="alignment">The minimum width of the formatted value.</param>
-        /// <param name="format">The format string to use.</param>
-        public void AppendFormatted(ReadOnlySpan<char> value, int alignment = 0, string? format = null) => AppendFormatted(value.ToString(), alignment, format);
-#pragma warning restore RS0027 // API with optional parameter(s) should have the most parameters amongst its public overloads
-#endif
-
-        // NOTE: All the overloads involving format and/or alignment are not super efficient.
-        // This code path is only for when an assert is failing, so that's not the common scenario
-        // and should be okay if not very optimized.
-        // A more efficient implementation that can be used for .NET 6 and later is to delegate the work to
-        // the BCL's StringBuilder.AppendInterpolatedStringHandler
-
-        /// <summary>Appends a formatted value to the interpolated message.</summary>
-        /// <typeparam name="T">The type of the value being appended.</typeparam>
-        /// <param name="value">The value to append.</param>
-        /// <param name="format">The format string to use.</param>
-        public void AppendFormatted<T>(T value, string? format) => _builder!.AppendFormat(null, $"{{0:{format}}}", value);
-
-        /// <summary>Appends a formatted value to the interpolated message.</summary>
-        /// <typeparam name="T">The type of the value being appended.</typeparam>
-        /// <param name="value">The value to append.</param>
-        /// <param name="alignment">The minimum width of the formatted value.</param>
-        public void AppendFormatted<T>(T value, int alignment) => _builder!.AppendFormat(null, $"{{0,{alignment}}}", value);
-
-        /// <summary>Appends a formatted value to the interpolated message.</summary>
-        /// <typeparam name="T">The type of the value being appended.</typeparam>
-        /// <param name="value">The value to append.</param>
-        /// <param name="alignment">The minimum width of the formatted value.</param>
-        /// <param name="format">The format string to use.</param>
-        public void AppendFormatted<T>(T value, int alignment, string? format) => _builder!.AppendFormat(null, $"{{0,{alignment}:{format}}}", value);
-
-        /// <summary>Appends a formatted value to the interpolated message.</summary>
-        /// <param name="value">The value to append.</param>
-        public void AppendFormatted(string? value) => _builder!.Append(value);
-
-#pragma warning disable RS0027 // API with optional parameter(s) should have the most parameters amongst its public overloads
-
-        /// <summary>Appends a formatted value to the interpolated message.</summary>
-        /// <param name="value">The value to append.</param>
-        /// <param name="alignment">The minimum width of the formatted value.</param>
-        /// <param name="format">The format string to use.</param>
-        public void AppendFormatted(string? value, int alignment = 0, string? format = null) => _builder!.AppendFormat(null, $"{{0,{alignment}:{format}}}", value);
-
-        /// <summary>Appends a formatted value to the interpolated message.</summary>
-        /// <param name="value">The value to append.</param>
-        /// <param name="alignment">The minimum width of the formatted value.</param>
-        /// <param name="format">The format string to use.</param>
-        public void AppendFormatted(object? value, int alignment = 0, string? format = null) => _builder!.AppendFormat(null, $"{{0,{alignment}:{format}}}", value);
-#pragma warning restore RS0027 // API with optional parameter(s) should have the most parameters amongst its public overloads
-    }
-
     /// <summary>
     /// Asserts that the delegate <paramref name="action"/> throws an exception of type <typeparamref name="TException"/>
     /// (or derived type) and throws <c>AssertFailedException</c> if code does not throw exception or throws
@@ -448,50 +184,6 @@ public sealed partial class Assert
         return message.ComputeAssertion(actionExpression);
     }
 
-    private static TException ThrowsException<TException>(Action action, bool isStrictType, string? message, string actionExpression, [CallerMemberName] string assertMethodName = "")
-        where TException : Exception
-    {
-        TelemetryCollector.TrackAssertionCall(GetTrackedThrowsName(assertMethodName));
-
-        _ = action ?? throw new ArgumentNullException(nameof(action));
-        _ = message ?? throw new ArgumentNullException(nameof(message));
-
-        ThrowsExceptionState state = IsThrowsFailing<TException>(action, isStrictType);
-        if (state.FailureKind != ThrowsFailureKind.NotFailing)
-        {
-            ReportThrowsFailed<TException>(isStrictType, state, message, actionExpression, assertMethodName);
-        }
-        else
-        {
-            return (TException)state.ExceptionThrown!;
-        }
-
-        // Reached when ReportThrowsFailed records the failure into the active AssertScope and returns instead of throwing.
-        return null!;
-    }
-
-    private static TException ThrowsException<TException>(Action action, bool isStrictType, Func<Exception?, string> messageBuilder, string actionExpression, [CallerMemberName] string assertMethodName = "")
-        where TException : Exception
-    {
-        TelemetryCollector.TrackAssertionCall(GetTrackedThrowsName(assertMethodName));
-
-        _ = action ?? throw new ArgumentNullException(nameof(action));
-        _ = messageBuilder ?? throw new ArgumentNullException(nameof(messageBuilder));
-
-        ThrowsExceptionState state = IsThrowsFailing<TException>(action, isStrictType);
-        if (state.FailureKind != ThrowsFailureKind.NotFailing)
-        {
-            ReportThrowsFailed<TException>(isStrictType, state, messageBuilder(state.ExceptionThrown), actionExpression, assertMethodName);
-        }
-        else
-        {
-            return (TException)state.ExceptionThrown!;
-        }
-
-        // Reached when ReportThrowsFailed records the failure into the active AssertScope and returns instead of throwing.
-        return null!;
-    }
-
     /// <summary>
     /// Asserts that the delegate <paramref name="action"/> throws an exception of type <typeparamref name="TException"/>
     /// (or derived type) and throws <c>AssertFailedException</c> if code does not throw exception or throws
@@ -607,226 +299,4 @@ public sealed partial class Assert
     public static Task<TException> ThrowsExactlyAsync<TException>(Func<Task> action, Func<Exception?, string> messageBuilder, [CallerArgumentExpression(nameof(action))] string actionExpression = "")
         where TException : Exception
         => ThrowsExceptionAsync<TException>(action, isStrictType: true, messageBuilder, actionExpression);
-
-    private static async Task<TException> ThrowsExceptionAsync<TException>(Func<Task> action, bool isStrictType, string? message, string actionExpression, [CallerMemberName] string assertMethodName = "")
-        where TException : Exception
-    {
-        TelemetryCollector.TrackAssertionCall(GetTrackedThrowsName(assertMethodName));
-
-        _ = action ?? throw new ArgumentNullException(nameof(action));
-        _ = message ?? throw new ArgumentNullException(nameof(message));
-
-        ThrowsExceptionState state = await IsThrowsAsyncFailingAsync<TException>(action, isStrictType).ConfigureAwait(false);
-        if (state.FailureKind != ThrowsFailureKind.NotFailing)
-        {
-            ReportThrowsFailed<TException>(isStrictType, state, message, actionExpression, assertMethodName);
-        }
-        else
-        {
-            return (TException)state.ExceptionThrown!;
-        }
-
-        // Reached when ReportThrowsFailed records the failure into the active AssertScope and returns instead of throwing.
-        return null!;
-    }
-
-    private static async Task<TException> ThrowsExceptionAsync<TException>(Func<Task> action, bool isStrictType, Func<Exception?, string> messageBuilder, string actionExpression, [CallerMemberName] string assertMethodName = "")
-        where TException : Exception
-    {
-        TelemetryCollector.TrackAssertionCall(GetTrackedThrowsName(assertMethodName));
-
-        _ = action ?? throw new ArgumentNullException(nameof(action));
-        _ = messageBuilder ?? throw new ArgumentNullException(nameof(messageBuilder));
-
-        ThrowsExceptionState state = await IsThrowsAsyncFailingAsync<TException>(action, isStrictType).ConfigureAwait(false);
-        if (state.FailureKind != ThrowsFailureKind.NotFailing)
-        {
-            ReportThrowsFailed<TException>(isStrictType, state, messageBuilder(state.ExceptionThrown), actionExpression, assertMethodName);
-        }
-        else
-        {
-            return (TException)state.ExceptionThrown!;
-        }
-
-        // Reached when ReportThrowsFailed records the failure into the active AssertScope and returns instead of throwing.
-        return null!;
-    }
-
-    [DebuggerDisableUserUnhandledExceptions]
-    private static async Task<ThrowsExceptionState> IsThrowsAsyncFailingAsync<TException>(Func<Task> action, bool isStrictType)
-        where TException : Exception
-    {
-        try
-        {
-            await action().ConfigureAwait(false);
-        }
-        catch (Exception ex)
-        {
-            bool isExceptionOfType = isStrictType
-                ? typeof(TException) == ex.GetType()
-                : ex is TException;
-
-            return isExceptionOfType
-                ? ThrowsExceptionState.CreateNotFailingState(ex)
-                : ThrowsExceptionState.CreateWrongTypeState(ex);
-        }
-
-        return ThrowsExceptionState.CreateNoExceptionState();
-    }
-
-    [DebuggerDisableUserUnhandledExceptions]
-    private static ThrowsExceptionState IsThrowsFailing<TException>(Action action, bool isStrictType)
-        where TException : Exception
-    {
-        try
-        {
-            action();
-        }
-        catch (Exception ex)
-        {
-            bool isExceptionOfType = isStrictType
-                ? typeof(TException) == ex.GetType()
-                : ex is TException;
-
-            return isExceptionOfType
-                ? ThrowsExceptionState.CreateNotFailingState(ex)
-                : ThrowsExceptionState.CreateWrongTypeState(ex);
-        }
-
-        return ThrowsExceptionState.CreateNoExceptionState();
-    }
-
-    [StackTraceHidden]
-    private static void ReportThrowsFailed<TException>(
-        bool isStrictType,
-        ThrowsExceptionState state,
-        string? userMessage,
-        string actionExpression,
-        string assertMethodName)
-        where TException : Exception
-    {
-        Type expectedType = typeof(TException);
-        string expectedTypeName = GetDisplayTypeName(expectedType, includeNamespace: false);
-        string expectedTypeFullName = GetDisplayTypeName(expectedType, includeNamespace: true);
-
-        StructuredAssertionMessage message;
-
-        if (state.FailureKind == ThrowsFailureKind.NoExceptionThrown)
-        {
-            string summary = isStrictType
-                ? $"Expected exception of exact type {expectedTypeName} but no exception was thrown."
-                : $"Expected exception of type {expectedTypeName} (or derived) but no exception was thrown.";
-
-            message = new StructuredAssertionMessage(summary)
-                .WithUserMessage(userMessage)
-                .WithCallSiteExpression(FormatCallSiteExpression($"Assert.{assertMethodName}<{expectedTypeName}>", actionExpression, "action"));
-        }
-        else
-        {
-            Exception actualException = state.ExceptionThrown!;
-            Type actualType = actualException.GetType();
-            string actualTypeName = GetDisplayTypeName(actualType, includeNamespace: false);
-            string actualTypeFullName = GetDisplayTypeName(actualType, includeNamespace: true);
-
-            string summary = isStrictType
-                ? $"Expected exception of exact type {expectedTypeName} but caught {actualTypeName}."
-                : $"Expected exception of type {expectedTypeName} (or derived) but caught {actualTypeName}.";
-
-            string expectedTypeLabel = isStrictType ? expectedTypeFullName : $"{expectedTypeFullName} (or derived)";
-
-            EvidenceBlock evidence = EvidenceBlock.Create()
-                .AddLine("expected type:", expectedTypeLabel)
-                .AddLine("actual type:", actualTypeFullName)
-                .AddLine("actual exception:", $"{actualTypeFullName}: {actualException.Message}");
-
-            message = new StructuredAssertionMessage(summary)
-                .WithUserMessage(userMessage)
-                .WithEvidence(evidence)
-                .WithCallSiteExpression(FormatCallSiteExpression($"Assert.{assertMethodName}<{expectedTypeName}>", actionExpression, "action"));
-        }
-
-        ReportAssertFailed(message);
-    }
-
-    // Renders a type name without the CLR backtick-arity suffix and with closed generic arguments expanded recursively
-    // (e.g. "MyException`1" with int → "MyException<Int32>") so it stays readable in summary lines and pasteable in call-site lines.
-    private static string GetDisplayTypeName(Type type, bool includeNamespace)
-    {
-        if (!type.IsGenericType)
-        {
-            return includeNamespace ? type.FullName ?? type.Name : type.Name;
-        }
-
-        string name = type.Name;
-        int tick = name.IndexOf('`');
-        if (tick >= 0)
-        {
-            name = name.Substring(0, tick);
-        }
-
-        if (includeNamespace && !string.IsNullOrEmpty(type.Namespace))
-        {
-            name = $"{type.Namespace}.{name}";
-        }
-
-        Type[] args = type.GetGenericArguments();
-        StringBuilder sb = new(name.Length + 16);
-        sb.Append(name);
-        sb.Append('<');
-        for (int i = 0; i < args.Length; i++)
-        {
-            if (i > 0)
-            {
-                sb.Append(", ");
-            }
-
-            sb.Append(GetDisplayTypeName(args[i], includeNamespace));
-        }
-
-        sb.Append('>');
-        return sb.ToString();
-    }
-
-    private enum ThrowsFailureKind : byte
-    {
-        NotFailing,
-        NoExceptionThrown,
-        WrongExceptionType,
-    }
-
-    [StackTraceHidden]
-    private readonly struct ThrowsExceptionState
-    {
-        public Exception? ExceptionThrown { get; }
-
-        public ThrowsFailureKind FailureKind { get; }
-
-        private ThrowsExceptionState(ThrowsFailureKind failureKind, Exception? exceptionThrown)
-        {
-            ExceptionThrown = exceptionThrown;
-            FailureKind = failureKind;
-        }
-
-        public static ThrowsExceptionState CreateWrongTypeState(Exception exceptionThrown)
-            => new(ThrowsFailureKind.WrongExceptionType, exceptionThrown);
-
-        public static ThrowsExceptionState CreateNoExceptionState()
-            => new(ThrowsFailureKind.NoExceptionThrown, null);
-
-        public static ThrowsExceptionState CreateNotFailingState(Exception exception)
-            => new(ThrowsFailureKind.NotFailing, exception);
-    }
-
-    // assertMethodName comes from [CallerMemberName] for the Throws/ThrowsExactly/ThrowsAsync/
-    // ThrowsExactlyAsync helpers — a small fixed set. Use a switch to avoid allocating a fresh
-    // "Assert." + name string on every call.
-    private static string GetTrackedThrowsName(string assertMethodName)
-        => assertMethodName switch
-        {
-            "Throws" => "Assert.Throws",
-            "ThrowsExactly" => "Assert.ThrowsExactly",
-            "ThrowsAsync" => "Assert.ThrowsAsync",
-            "ThrowsExactlyAsync" => "Assert.ThrowsExactlyAsync",
-            _ => string.Concat("Assert.", assertMethodName),
-        };
 }
